More renames, move aws/ssm/index.ts to ssh/index.ts
